# cmd/api/docs/swagger.yaml
definitions:
  github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.Product:
    properties:
      description:
        minLength: 15
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    required:
    - description
    - name
    - price
    - stock
    type: object
  github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.User:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: fazilkp2000@gmail.com
    name: API Support
    url: https://fazilnbr.github.io/mypeosolal.web.portfolio/
  description: This is a sample server Job Portal server. You can visit the GitHub
    repository at https://github.com/fazilnbr/Job_Portal_Project
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Workey API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      operationId: authentication login
      parameters:
      - description: auth login
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      summary: authentication login
      tags:
      - Authentication
  /auth/register:
    post:
      operationId: SignUp authentication
      parameters:
      - description: Worker Login
        in: body
        name: WorkerLogin
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.User'
          - properties:
              password:
                type: string
              username:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      summary: SignUp for users
      tags:
      - User Authentication
      - User Authentication
  /auth/token-refresh:
    get:
      operationId: Refresh access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Refresh The Access Token
      tags:
      - Refresh Token
  /product:
    delete:
      operationId: deleteproduct
      parameters:
      - description: 'Product Id : '
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a product by ID
      tags:
      - Product
    get:
      operationId: retriveproduct
      parameters:
      - description: 'Product Id : '
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Retrieve a product by ID
      tags:
      - Product
    post:
      operationId: createproduct
      parameters:
      - description: Product Detials
        in: body
        name: productdetials
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Product
    put:
      operationId: updateproduct
      parameters:
      - description: Product Detials
        in: body
        name: productdetials
        required: true
        schema:
          $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Update an existing product
      tags:
      - Product
  /product/all:
    get:
      operationId: listproduct
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fazilnbr_GoCart-grpc-API-Gateway_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: List all available products
      tags:
      - Product
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"